refactor: use base class to remove unneeded CSS
